+++ pkg/resource/logging-secret/logging-secret.go
@@ -47,7 +47,7 @@ func GenerateLoggingSecret(lc loggedcluster.Interface, loggingCredentialsSecret 
 func SecretMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
 	metadata := metav1.ObjectMeta{
 		Name:      GetLoggingSecretName(lc),
-		Namespace: lc.GetAppsNamespace(),
+		Namespace: lc.GetNamespace(),
 		Labels:    map[string]string{},
 	}
 
@@ -56,5 +56,5 @@ func SecretMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
 }
 
 func GetLoggingSecretName(lc loggedcluster.Interface) string {
-	return fmt.Sprintf("%s-%s", lc.GetClusterName(), loggingClientSecretName)
+	return fmt.Sprintf("%s-%s", lc.GetName(), loggingClientSecretName)
 }